projects: eval-adxl313z: use the example infrastructure
Migrate eval-adxl313z to the shared examples infrastructure (examples.mk):
replace the Makefile BASIC_EXAMPLE/IIO_EXAMPLE toggles with EXAMPLE := iio
+ include examples.mk, split the iio sources into
src/examples/iio/example.mk, drop examples_src.mk and platform_includes.h,
and shorten the iio example directory (iio_example -> iio). Both platform
main.c files keep their device/platform init (spi_init, and the stm32
clock + stm32_init) and defer to example_main(); the basic example owns
its UART stdio init and the iio example includes iio_app.h directly.
Switch builds.json flags to EXAMPLE= and update the README to select the
example with make EXAMPLE=<name>.

While here, fix a latent bug in the maxim parameters.h that left
UART_IRQ_ID defined only under IIO_SUPPORT, so the basic example failed to
compile on maxim (common_data.c references it unconditionally). builds.json
only exercised stm32, so this was never caught.
