+++ projects/eval-adxl313z/builds.json
@@ -1,10 +1,10 @@
 {
   "stm32": {
-    "basic_example": {
-      "flags" : "BASIC_EXAMPLE=y IIO_EXAMPLE=n"
+    "basic": {
+      "flags" : "EXAMPLE=basic"
     },
-	"iio_example": {
-      "flags" : "BASIC_EXAMPLE=n IIO_EXAMPLE=y"
+    "iio": {
+      "flags" : "EXAMPLE=iio"
     }
   }
 }
