+++ projects/eval-adxl313z/src/examples/basic/basic_example.c
@@ -31,9 +31,9 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
 
-#include "basic_example.h"
 #include "common_data.h"
 #include "adxl313.h"
+#include "no_os_uart.h"
 #include "no_os_delay.h"
 #include "no_os_print_log.h"
 #include "no_os_error.h"
@@ -44,10 +44,17 @@
  * @return ret - Result of the example execution. If working correctly, will
  *               execute continuously the while(1) loop and will not return.
 *******************************************************************************/
-int basic_example_main()
+int example_main()
 {
 	struct adxl313_dev *adxl313;
+	struct no_os_uart_desc *uart_desc;
 	int ret = -EINVAL;
+
+	ret = no_os_uart_init(&uart_desc, &uip);
+	if (ret)
+		return ret;
+
+	no_os_uart_stdio(uart_desc);
 
 	/* Clear the screen. */
 	printf("%c", 27);
