--- projects/eval-adxl313z/src/examples/examples_src.mk
@@ -1,22 +1,0 @@
-ifeq (y,$(strip $(IIO_EXAMPLE)))
-IIOD=y
-CFLAGS += -DIIO_EXAMPLE
-SRCS += $(PROJECT)/src/examples/iio_example/iio_example.c
-INCS += $(PROJECT)/src/examples/iio_example/iio_example.h
-endif
-
-ifeq (y,$(strip $(BASIC_EXAMPLE)))
-CFLAGS += -DBASIC_EXAMPLE
-SRCS += $(PROJECT)/src/examples/basic/basic_example.c
-INCS += $(PROJECT)/src/examples/basic/basic_example.h
-endif
-
-ifeq (y,$(strip $(IIOD)))
-SRC_DIRS += $(NO-OS)/iio/iio_app
-INCS += $(DRIVERS)/accel/adxl313/iio_adxl313.h
-
-SRCS += $(DRIVERS)/accel/adxl313/iio_adxl313.c
-
-INCS += $(INCLUDE)/no_os_list.h \
-		$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.h
-endif
