+++ projects/eval-adxl313z/src/examples/iio/example.mk
@@ -1,0 +1,4 @@
+IIOD = y
+
+INCS += $(DRIVERS)/accel/adxl313/iio_adxl313.h
+SRCS += $(DRIVERS)/accel/adxl313/iio_adxl313.c
