+++ projects/eval-adxl313z/src/platform/maxim/parameters.h
@@ -42,9 +42,9 @@
 
 #ifdef IIO_SUPPORT
 #define INTC_DEVICE_ID  0
-#define UART_IRQ_ID     UART0_IRQn
 #endif
 
+#define UART_IRQ_ID     UART0_IRQn
 #define UART_DEVICE_ID  0
 #define UART_BAUDRATE   57600
 #define UART_EXTRA      &xuip
